requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
WTForms==3.1.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
email-validator==2.1.0
dnspython==2.4.2
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
reportlab==4.0.7
Pillow==10.1.0
publicsuffix2==2.20191221
pytest==7.4.3
pytest-flask==1.3.0
